Add whips information to role presenter
Adding the information about the given whip organisation for ministerial
roles that are whips.

Part of the work to migrate rendering of the ministers index page
to collections.

https://trello.com/c/loAlkIp7/2063-8-migrate-government-ministers-to-collections

The purpose of these changes is to allow collections to sort the list
of whips at a rendering level. This will eventually allow some of the
function of the `ministerial_roles_controller` to be removed out to
collections.

https://github.com/alphagov/whitehall/blob/1512033670b30512872b56fbf4801cb438be55ec/app/controllers/ministerial_roles_controller.rb#L43-L60