Add a migration
Add a migration to add a string value to edition for for the 
image_display_type. With the possible values being "organisation_image", 
"no_image" or "custom_image".
(we considered using a boolean however when we tested it out we could 
not set the value to nil with the radio buttons since it was treated as 
false)
TODO: We intend on spliting this into a migration for the column and a 
data_migration to update the values in line with this documentation: 
https://github.com/alphagov/whitehall/tree/master/db/data_migration#data-migrations