Add radio buttons to set image_display_type
Add radio buttons to allow the user to choose what image will be used on 
the edition. Show and hide image uploader depending on the choice. 
TODO: This currently will show on all editions which use the 
image_fields however we only what to show this for case_studies, we plan 
on adding some logic to only display the options onm case_studies.